imprv(code quality): remove dead code path
Code paths are sometimes hard to reason about, but months of runtime
and log inspection confirm that exceptions here are all raised in
Contact.update_info() in which InvalidUserException are caught and
turned into "item-not-found" XMPPErrors.

This is a bit messy… but removing a dead code path makes it a bit
better, hopefully.